Move CodeGenerator to codegen package, fixed a typo.
