+++ trans/src/wich/parser/Wich.g4
@@ -24,6 +24,7 @@ SOFTWARE.
 
 grammar Wich;
 
+
 file : script ;
 
 script : (statement | function)+ EOF ;
