+++ trans/test/TestSymbolDefine.java
@@ -23,7 +23,6 @@ SOFTWARE.
 */
 
 import org.antlr.symtab.GlobalScope;
-import org.junit.Before;
 import org.junit.Test;
 
 import static org.junit.Assert.assertEquals;
